infrastructure/requirements.txt
-e .
aws-cdk-lib>=2.8.0
constructs>=10.0.59